Add an optional flag to all handlers to allow raw status return values
As proposed in issue #319, this introduces an optional status_prefix boolean that can be added to any resource handler definition. If set to True (or omitted), status returned by the handler is still placed beneath the handler_id in the resulting object's status subresource. If set to False, the status returned by the handler is use at the top-level of the status subresource.

Default handler with status_prefix=True:
  @kopf.on.create('zalando.org', 'v1', 'kopfexamples')
  def create(...):
    return {'my': 'result}

Result:
  ...
  Status:
    Create:
      My: Result

Handler with status_prefix=False:
  @kopf.on.create('zalando.org', 'v1', 'kopfexamples', status_prefix=False)
  def create(...):
    return {'my': 'result}

Result:
  ...
  Status:
    My: Result